requirements.txt
colormath
python-slugify
